feat: standardize header configuration
Most screen types supported one or both of the "stop name" and "stop ID"
header configuration styles, and there was some duplication in how these
were encoded and decoded. This change standardizes things to an extent:

* 🔶 Bus E-ink now supports `stop_name`.
* ☑️ Bus Shelter already supported both types and is unchanged.
* 🔶 Busway (Sectional) now supports `stop_id`.
* ☑️ DUP already supported both types and is unchanged.
* 🚫 Elevator is unchanged (no header configuration).
* 🚫 GL E-ink is unchanged (has its own unique header type).
* 🚫 On-Bus is unchanged (no header configuration).
* 🔶 Pre-Fare now supports `stop_name`.

This change is backwards-compatible with the existing format.
